refactor(executors): 拆分共享运行时职责
