+++ src/executors/claude-sdk-trace.ts
@@ -1,5 +1,5 @@
 import type { ToolCallInfo, TurnInfo } from '../types/index.js';
-import type { ClaudeSdkBaseMessage } from './shared.js';
+import type { ClaudeSdkBaseMessage } from './claude-protocol.js';
 import { safeSliceForJson } from '../util/safe-slice.js';
 import { isToolResultFailureText } from '../observability/text-signals.js';
 import { normalizeToolIdentity } from '../shared/tool-identity.js';
